more work on factories, break fake_agency out
