+++ acquisitions/factories.py
@@ -1,37 +1,6 @@
 import factory
-import random
-from faker import Faker
-from faker.providers import BaseProvider
 from acquisitions import models
-
-fake = Faker()
-
-
-class AgencyProvider(BaseProvider):
-    agency_parts = (
-        (
-            "Department of", "Office of", "Bureau of"
-        ),
-        (
-            "the Interior", "Administrating", "Hats", "Management", "Labor",
-            "Finance", "Departments", "Flying"
-        )
-    )
-
-    extra_parts = (
-        "Synergy", "Failure", "High-Profile Success", "First Aid", "Gravy",
-        "Sandwiches", "Wine", "Budget", "Style"
-    )
-
-    def agency(self):
-        result = []
-        for part in self.agency_parts:
-            result.append(self.random_element(part))
-        if random.randint(0, 100) > 70:
-            result.append("and")
-            result.append(self.random_element(self.extra_parts))
-
-        return " ".join(result)
+from acquisitions.providers.fake_agency import AgencyProvider
 
 factory.Faker.add_provider(AgencyProvider)
 
@@ -49,6 +18,18 @@ class SubagencyFactory(factory.django.DjangoModelFactory):
 
     name = factory.Faker("agency")
     agency = factory.SubFactory(AgencyFactory)
+
+
+class ContractingOfficeFactory(factory.django.DjangoModelFactory):
+    pass
+
+
+class ContractingOfficerFactory(factory.django.DjangoModelFactory):
+    pass
+
+
+class CORFactory(factory.django.DjangoModelFactory):
+    pass
 
 
 class TrackFactory(factory.django.DjangoModelFactory):
@@ -73,5 +54,21 @@ class ActorFactory(factory.django.DjangoModelFactory):
     name = factory.Faker("company")
 
 
+class VendorFactory(factory.django.DjangoModelFactory):
+    pass
+
+
+class RoleFactory(factory.django.DjangoModelFactory):
+    pass
+
+
 class AcquisitionFactory(factory.django.DjangoModelFactory):
+    pass
+
+
+class EvaluatorFactory(factory.django.DjangoModelFactory):
+    pass
+
+
+class ReleaseFactory(factory.django.DjangoModelFactory):
     pass
