+++ acquisitions/management/commands/seed_database.py
@@ -6,4 +6,5 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         factories.ActorFactory.create_batch(5)
-        factories.AgencyFactory.create_batch(5)
+        # factories.AgencyFactory.create_batch(5)
+        factories.SubagencyFactory.create_batch(5)
